+++ core/lib/refinery/plugins.rb
@@ -11,11 +11,6 @@ module Refinery
 
     def_delegators :@plugins, :<<, :each, :delete_if
 
-    def find_by_model(model)
-      model = model.constantize if model.is_a? String
-      detect { |plugin| plugin.activity.any? {|activity| activity.klass == model } }
-    end
-
     def find_by_name(name)
       detect { |plugin| plugin.name == name }
     end