+++ dashboard/spec/features/refinery/admin/dashboard_spec.rb
@@ -30,31 +30,4 @@ describe "dashboard" do
       end
     end
   end
-
-  describe "latest activity" do
-    before do
-      3.times { |n| FactoryGirl.create :refinery_user, :username => "ugisozols#{n}" }
-      3.times { |n| FactoryGirl.create :page, :title => "Refinery CMS #{n}" }
-    end
-
-    it "shows created tracked objects" do
-      visit refinery.admin_dashboard_path
-
-      page.should have_content("Latest Activity")
-      3.times { |n| page.should have_content("Ugisozols#{n} user was added") }
-      3.times { |n| page.should have_content(/Refinery cms #{n} page was (added|updated)/) }
-    end
-
-    # see https://github.com/refinery/refinerycms/issues/1673
-    it "uses proper link for nested pages" do
-      # we need to increase updated_at because dashboard entries are sorted by
-      # updated_at column and we need this page to be at the top of the list
-      nested = FactoryGirl.create(:page, :parent_id => Refinery::Page.last.id,
-                                         :updated_at => Time.now + 10.seconds)
-
-      visit refinery.admin_dashboard_path
-
-      page.should have_selector("a[href='#{refinery.edit_admin_page_path(nested.nested_url)}']")
-    end
-  end
 end