+++ pages/lib/refinery/pages/engine.rb
@@ -23,11 +23,6 @@ module Refinery
           plugin.pathname = root
           plugin.name = 'refinery_pages'
           plugin.menu_match = %r{refinery/page(_part|s_dialog)?s(/preview)?$}
-          plugin.activity = {
-            :class_name => :'refinery/page',
-            :nested_with => [:nested_url],
-            :use_record_in_nesting => false
-          }
           plugin.url = proc { Refinery::Core::Engine.routes.url_helpers.admin_pages_path }
         end
       end