+++ changelog.md
@@ -7,6 +7,7 @@
 * Removed Pages#cache_pages_backend. [#2375](https://github.com/refinery/refinerycms/pull/2375) [Philip Arndt](https://github.com/parndt)
 * Updated how `_make_sortable` works to take an options hash, requiring manual file changes. [Philip Arndt](https://github.com/parndt)
 * Removed `attr_accessible` in favour of strong parameters. [Philip Arndt](https://github.com/parndt) [#2518](https://github.com/refinery/refinerycms/pull/2518)
+* Removed the latest activity feed from the Dashboard. [Philip Arndt](https://github.com/parndt) [2634](https://github.com/refinery/refinerycms/pull/2633)
 * [See full list](https://github.com/refinery/refinerycms/compare/2-1-stable...master)
 
 ## 2.1.2 [14 March 2014]